+++ packages/antd/src/templates/ErrorList/index.tsx
@@ -8,6 +8,7 @@ import {
   FormContextType,
   RJSFSchema,
   StrictRJSFSchema,
+  TranslatableString,
 } from "@rjsf/utils";
 
 /** The `ErrorList` component is the template that renders the all the errors associated with the fields in the `Form`
@@ -18,7 +19,8 @@ export default function ErrorList<
   T = any,
   S extends StrictRJSFSchema = RJSFSchema,
   F extends FormContextType = any
->({ errors }: ErrorListProps<T, S, F>) {
+>({ errors, registry }: ErrorListProps<T, S, F>) {
+  const { translateString } = registry;
   const renderErrors = () => (
     <List className="list-group" size="small">
       {errors.map((error, index) => (
@@ -36,7 +38,7 @@ export default function ErrorList<
     <Alert
       className="panel panel-danger errors"
       description={renderErrors()}
-      message="Errors"
+      message={translateString(TranslatableString.ErrorsLabel)}
       type="error"
     />
   );