+++ packages/semantic-ui/src/ErrorList/ErrorList.tsx
@@ -5,6 +5,7 @@ import {
   FormContextType,
   RJSFSchema,
   StrictRJSFSchema,
+  TranslatableString,
 } from "@rjsf/utils";
 
 /** The `ErrorList` component is the template that renders the all the errors associated with the fields in the `Form`
@@ -15,10 +16,13 @@ export default function ErrorList<
   T = any,
   S extends StrictRJSFSchema = RJSFSchema,
   F extends FormContextType = any
->({ errors }: ErrorListProps<T, S, F>) {
+>({ errors, registry }: ErrorListProps<T, S, F>) {
+  const { translateString } = registry;
   return (
     <Message negative>
-      <Message.Header>Errors</Message.Header>
+      <Message.Header>
+        {translateString(TranslatableString.ErrorsLabel)}
+      </Message.Header>
       <Message.List>
         {errors.map((error, index) => (
           <Message.Item key={`error-${index}`}>{error.stack}</Message.Item>